requirements.txt
# AutoPBR Pipeline Dependencies

# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
open-clip-torch>=2.20.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Geospatial Processing
shapely>=2.0.0
geopandas>=0.13.0
pyproj>=3.6.0
Rtree>=1.0.0

# Configuration & Utilities
PyYAML>=6.0
requests>=2.31.0
